refactor(app): Refactor API
