+++ src/index.ts
@@ -1,1 +1,2 @@
 export * from './lib/app';
+export { TDispatch } from './lib/dispatch';
