+++ src/lib/app.ts
@@ -1,30 +1,14 @@
-import yaml from 'js-yaml';
+import { TConfig, TPatchedManifest, TRawManifest } from '../types';
 
-import {
-  EAccept,
-  EOperation,
-  TConfig,
-  TContext,
-  TDispatchFn,
-  TMessage,
-  TPatchedManifest,
-  TRawManifest,
-} from '../types';
-
+import { createAgent, TAgent } from './agent';
 import { validateConfig } from './config';
-import { makeContext } from './context';
-import {
-  createAgent,
-  createServer,
-  startServer,
-  TAgent,
-  TServer,
-} from './http';
-import { ensureManifest, patchManifest, validateManifest } from './manifest';
+import { createContext } from './context';
+import { createDispatch } from './dispatch';
+import { createServer, startServer, TServer } from './http';
+import { patchManifest, syncManifest, validateManifest } from './manifest';
 
 export type TApp = {
   readonly httpServer: TServer;
-  readonly context: TContext;
   readonly agent: TAgent;
   readonly config: TConfig;
   readonly patchedManifest: TPatchedManifest;
@@ -43,16 +27,25 @@ export const createApp = (config: TConfig, manifest: TRawManifest): TApp => {
     throw new Error(manifestValidation.error);
   }
 
-  const agent = createAgent(config);
-  const context = makeContext(agent);
+  const agent = createAgent({
+    baseURL: config.AIDBOX_URL,
+    auth: {
+      username: config.AIDBOX_CLIENT_ID,
+      password: config.AIDBOX_CLIENT_SECRET,
+    },
+  });
   const { subscriptionHandlers, patchedManifest } = patchManifest(manifest);
-  const httpServer = createServer(
-    dispatch(config, patchedManifest, context, subscriptionHandlers)
+  const context = createContext(agent);
+  const dispatch = createDispatch(
+    config,
+    patchedManifest,
+    context,
+    subscriptionHandlers
   );
+  const httpServer = createServer(dispatch);
 
   return {
     httpServer,
-    context,
     agent,
     config,
     patchedManifest,
@@ -60,123 +53,21 @@ export const createApp = (config: TConfig, manifest: TRawManifest): TApp => {
 };
 
 export const startApp = async (app: TApp) => {
-  // eslint-disable-next-line functional/no-let
-  let isReady;
+  const { agent, config, patchedManifest, httpServer } = app;
   try {
-    isReady = await app.context.request(
-      {
-        url: '/__healthcheck',
-      },
-      false
-    );
-  } catch (e) {
-    console.log(e, 'error');
+    await waitForAidboxServer(agent);
+  } catch (err) {
+    console.error(`Aidbox server is unreachable.`);
+    process.exit(1);
+    return;
   }
-  if (!isReady) {
-    console.log('aidbox not ready', 'error');
-    process.exit(0);
-  }
-
-  await ensureManifest(app.agent, app.config, app.patchedManifest);
-  await startServer(app.httpServer);
+  await syncManifest(agent, config, patchedManifest);
+  await startServer(httpServer);
 };
 
-const resolveContentType = (msg: TMessage) => {
-  switch (msg.request?.headers?.accept) {
-    case EAccept.YAML:
-    case EAccept.TEXT:
-      return msg.request.headers.accept;
-    default:
-      return EAccept.JSON;
-  }
-};
-
-const checkAuthHeader = (
-  appId: string,
-  appSecret: string,
-  authHeader?: string
-) => {
-  if (!authHeader) {
-    return false;
-  }
-  const [auth] = authHeader.split(' ').slice(1, 2);
-  return auth === Buffer.from(`${appId}:${appSecret}`).toString('base64');
-};
-
-const dispatch: TDispatchFn = (
-  config,
-  manifest,
-  context,
-  subscriptionHandlers
-) => (req, res) => {
-  const sendResponse = (
-    text: string,
-    status = 200,
-    headers: Record<string, string> = {}
-  ) => {
-    // eslint-disable-next-line functional/immutable-data
-    res.statusCode = status;
-    Object.keys(headers).forEach((k: string) => {
-      res.setHeader(k, headers[k]);
-    });
-    res.end(text);
-  };
-
-  // eslint-disable-next-line functional/no-let
-  let reqBody = '';
-  req.on('data', (chunk) => {
-    reqBody += chunk.toString();
-  });
-  req.on('end', async () => {
-    console.log(reqBody);
-    try {
-      const msg = JSON.parse(reqBody) as TMessage;
-      res.setHeader('Content-Type', resolveContentType(msg));
-
-      if (
-        !checkAuthHeader(
-          config.APP_ID,
-          config.APP_SECRET,
-          req.headers.authorization
-        )
-      ) {
-        sendResponse(JSON.stringify({ message: 'Access Denied' }), 403);
-        return;
-      }
-
-      const operation = msg.type;
-
-      if (operation === EOperation.SUBSCRIPTION) {
-        const handlerId = msg.handler as string;
-        if (subscriptionHandlers[handlerId]) {
-          subscriptionHandlers[handlerId](context, msg);
-          sendResponse(JSON.stringify({ status: 'start subs' }));
-        }
-        return;
-      }
-
-      const operationId = msg.operation?.id;
-      if (
-        operation === EOperation.OPERATION &&
-        operationId &&
-        manifest.operations[operationId]
-      ) {
-        const { handler } = manifest.operations[operationId];
-        const { status, headers, resource, body } = await handler(context, msg);
-        if (msg.request.headers.accept === 'text/yaml') {
-          sendResponse(yaml.dump(resource, { noRefs: true }), status, headers);
-          return;
-        } else if (body) {
-          sendResponse(body, status, headers);
-          return;
-        } else {
-          sendResponse(JSON.stringify(resource), status, headers);
-          return;
-        }
-      }
-    } catch (e) {
-      console.log(e);
-      sendResponse(JSON.stringify(e));
-    }
+const waitForAidboxServer = async (agent: TAgent) => {
+  await agent.request({
+    url: '/__healthcheck',
+    responseType: 'text',
   });
 };
